refactor: update ElementController namespace and fix conditional logic in AdminLayout
