+++ app/Http/Controllers/Api/Admin/ElementController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Api\Admin;
 
+use App\Http\Controllers\Controller;
 use App\Models\Element;
 use Illuminate\Http\Request;
